[WIP] Add state aware IP allocation support
This commit is adding the -p option which can be used to provide
the result of a previous allocation in json format as input for
a new allocation, together with the main input file. The old
allocation is used to check if new entries are added compared to
the old ones (the node name and key names are used as unique keys)
and also to understand if the new allocation is different from the
old one (e.g. new subnets are used or a requested subnet is different
etc).